test_repo/requirements.txt
# Vulnerable dependencies for testing
django==2.0.0
flask==0.12.0
requests==2.6.0
pyyaml==3.12
pillow==5.0.0
urllib3==1.24.0
jinja2==2.10.0
cryptography==2.0
